1, 修改Model读表错误问题；2，测试队列（有问题，可能需要改成redis队列）；3，修改数据库名称以及创建

+++ tasks/FeedTask.php
@@ -14,35 +14,50 @@ class FeedTask extends \Phalcon\CLI\Task {
     private function _processQueue() {
         $di = $this->getDI();
 
+        $queue_key = \Util\ReadConfig::get('queue_keys.allfeeds', $di);
         $queue = \Util\BStalkClient::getInstance($di);
-        $queue->choose(\Util\ReadConfig::get('queue_keys.allfeeds', $di));
+        $queue->choose($queue_key);
+        $queue->watch($queue_key);
 
         $redis = \Util\RedisClient::getInstance($di);
         $cache_app_id_feeds = \Util\ReadConfig::get('redis_cache_keys.app_id_feeds', $di);
 
         $model = new FeedModel($this->getDI());
 
-        while(($job = $queue->peekReady()) !== false) {
-            $message = $job->getBody();
-            $oldMessage = $message;
-            $newMessage = msgpack_unpack($message);
-            $feed_id = $model->create($newMessage);
-            if($feed_id > 0) {
-                $key = sprintf($cache_app_id_feeds, $newMessage['app_id']);
-                $redis->zadd($key, -$newMessage['create_at'], $oldMessage);
+        while(true) {
+            if(false !== $queue->peekReady()) {
+                $job = $queue->reserve();
+                $message = $job->getBody();
+                $oldMessage = $message;
+                var_dump($message);
+                $newMessage = msgpack_unpack($message);
+                $feed_id = $model->create($newMessage);
 
-                $model->push($newMessage['app_id'],
-                    $newMessage['author_id'],
-                    $feed_id,
-                    $newMessage['create_at']);
+                if($feed_id > 0) {
+                    $key = sprintf($cache_app_id_feeds, $newMessage['app_id']);
+                    $job->delete();
+                    $redis->zadd($key, -$newMessage['create_at'], $oldMessage);
 
-                if($redis->zcard($key) > 1000) {
-                    $redis->zremrangebyrank($key, 501, -1);
+                    $model->push($newMessage['app_id'],
+                        $newMessage['author_id'],
+                        $feed_id,
+                        $newMessage['create_at']);
+
+                    if($redis->zcard($key) > 1000) {
+                        $redis->zremrangebyrank($key, 501, -1);
+                    }
+
+                    $job->delete();
+
                 }
 
-                $job->delete();
             }
+
+            exit;
+//            sleep(1);
         }
+
+
     }
